requirements.txt
uvicorn
fastapi
loguru
pydantic
torch
modelscope
transformers
pathlib
pdfminer-six
magic_pdf